Added early interface to the module manager, moved pid definition to arch-agnostic code

+++ kernel/include/pid.h
@@ -21,7 +21,7 @@
 
 #include <hack_stdint.h>
 
-typedef uint32_t PID;
+typedef size_t PID;
 const PID PID_INVALID = 0;
 const PID PID_KERNEL = 1;
